Improvements to Linux build scripts

+++ build-scripts/linux-x86_64/build_basilisk.sh
@@ -18,9 +18,11 @@ fi
 # If argument is provided then container will update before build
 UPDATEFIRST=$1
 
-cp mozconfigs/linux/x86_64/gtk3_unofficial_branding.mozconfig .mozconfig
+if [ ! -f .mozconfig ]; then
+    cp mozconfigs/linux/x86_64/gtk3_unofficial_branding.mozconfig .mozconfig
+fi
 
 echo "Starting Container..."
 
-# Slackware was selected because of the combination of stability and ease of use
-docker run -i -v $PWD:/share --rm -e UPDATEFIRST=$UPDATEFIRST -e UID=$(id -u) -e GID=$(id -g) -e USERNAME=$(whoami) -t aclemons/slackware:15.0-full /share/build-scripts/linux-x86_64/build_basilisk_subscripts/run_inside_docker.sh
+# Slackware was selected because of the combination of stability and ease of use. Slackware 15 was specifically chosen because it comes with GCC 11.
+docker run -i -v $PWD:/share --rm -e UPDATEFIRST=$UPDATEFIRST -e UID=$(id -u) -e GID=$(id -g) -e USERNAME=$(whoami) -e GROUPNAME=$(id -gn) -t aclemons/slackware:15.0-full /share/build-scripts/linux-x86_64/build_basilisk_subscripts/run_inside_docker.sh
